+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Version 0.4.0-dev (changes since 0.4.0 release go here)
 
+### Features
+* Added `to_regex` method to `skbio.sequence._iupac_sequence` ABC - it returns a regex object that matches all non-degenerate versions of the sequence.
+
 ## Version 0.4.0 (2015-07-08)
 
 Initial beta release. In addition to the changes detailed below, the following